Reset happy path UAT specs
Renamed file: easier to see Support and Delegate user specs together.
Reworked context to follow a better top-down decomposition
Added a few happy paths
